+++ NEWS.md
@@ -1,3 +1,8 @@
+# teamcolors 0.0.3
+
+* Added NWSL colors (thanks `adror1`!)
+* Added `show_sport_col()` function
+
 # teamcolors 0.0.2
 
 * Fixed Los Angeles of Anaheim division